testsite/requirements.txt
Django==2.2.27
django-extra-views==0.13.0
djangorestframework==3.12.4

# testsite-only
django-debug-toolbar==2.1
django-extensions==3.0.2
djaodjin-rules==2.5

# development
pylint==2.5.3
pylint-django==2.0.15
Sphinx==3.1.2
sphinxcontrib-httpdomain==1.7.0
